Move reset password page store to central auth store
