--- src/containers/ResetPasswordPage/constants.js
@@ -1,5 +1,0 @@
-export const RESET_PASSWORD = '[Reset Password]';
-
-export const RESET_PASSWORD_REQUEST = `${RESET_PASSWORD} RESET_PASSWORD_REQUEST`;
-export const RESET_PASSWORD_SUCCESS = `${RESET_PASSWORD} RESET_PASSWORD_SUCCESS`;
-export const RESET_PASSWORD_ERROR = `${RESET_PASSWORD} RESET_PASSWORD_ERROR`;
