+++ src/store/auth/actionTypes.js
@@ -15,3 +15,6 @@ export const FORGOT_PASSWORD_ERROR = `${AUTH} FORGOT_PASSWORD_ERROR`;
 export const REGISTER_REQUEST = `${AUTH} REGISTER_REQUEST`;
 export const REGISTER_SUCCESS = `${AUTH} REGISTER_SUCCESS`;
 export const REGISTER_ERROR = `${AUTH} REGISTER_ERROR`;
+export const RESET_PASSWORD_REQUEST = `${AUTH} RESET_PASSWORD_REQUEST`;
+export const RESET_PASSWORD_SUCCESS = `${AUTH} RESET_PASSWORD_SUCCESS`;
+export const RESET_PASSWORD_ERROR = `${AUTH} RESET_PASSWORD_ERROR`;
