+++ src/store/auth/actions.js
@@ -14,6 +14,9 @@ import {
   REGISTER_REQUEST,
   REGISTER_SUCCESS,
   REGISTER_ERROR,
+  RESET_PASSWORD_REQUEST,
+  RESET_PASSWORD_SUCCESS,
+  RESET_PASSWORD_ERROR,
 } from './actionTypes';
 
 export function login(email, password, setErrors) {
@@ -122,5 +125,28 @@ export function registerSuccess() {
 export function registerError() {
   return {
     type: REGISTER_ERROR,
+  };
+}
+
+export function resetPassword(password, passwordConfirmation, setErrors) {
+  return {
+    type: RESET_PASSWORD_REQUEST,
+    password,
+    passwordConfirmation,
+    meta: {
+      setErrors,
+    },
+  };
+}
+
+export function resetPasswordSuccess() {
+  return {
+    type: RESET_PASSWORD_SUCCESS,
+  };
+}
+
+export function resetPasswordError() {
+  return {
+    type: RESET_PASSWORD_ERROR,
   };
 }
