+++ src/store/auth/sagas.js
@@ -1,4 +1,4 @@
-import { takeLatest, call, put } from 'redux-saga/effects';
+import { takeLatest, call, put, select } from 'redux-saga/effects';
 import { push } from 'connected-react-router';
 import { setItem, removeItem } from 'utils/localStorage';
 import request from 'utils/request';
@@ -13,6 +13,8 @@ import {
   forgotPasswordError,
   registerSuccess,
   registerError,
+  resetPasswordSuccess,
+  resetPasswordError,
 } from './actions';
 import { startAction, stopAction } from '../loading/actions';
 import { enqueueSnackbar } from '../notifier/actions';
@@ -22,11 +24,15 @@ import {
   LOGOUT_REQUEST,
   FORGOT_PASSWORD_REQUEST,
   REGISTER_REQUEST,
+  RESET_PASSWORD_REQUEST,
 } from './actionTypes';
-import { DASHBOARD } from 'routes';
+import { DASHBOARD, LOGIN } from 'routes';
 import messages from 'containers/LoginPage/messages';
 import forgotPasswordMessages from 'containers/ForgotPasswordPage/messages';
+import resetPasswordMessages from 'containers/ResetPasswordPage/messages';
 import parseApiErrorsToFormik from 'utils/parseApiErrorsToFormik';
+
+const getRouterLocationSearch = (state) => state.router.location.search;
 
 export function* authorize({ type, email, password }) {
   try {
@@ -145,10 +151,46 @@ export function* register({
   }
 }
 
+export function* resetPassword({
+  type,
+  password,
+  passwordConfirmation,
+  meta: { setErrors },
+}) {
+  try {
+    yield put(startAction(type));
+
+    const search = yield select(getRouterLocationSearch);
+    const params = new URLSearchParams(search);
+    const token = params.get('forgot_password_token');
+    yield call(request, {
+      url: '/user/reset-password',
+      method: 'post',
+      data: { token, password, password_confirmation: passwordConfirmation },
+    });
+    yield put(resetPasswordSuccess());
+    yield put(
+      enqueueSnackbar({
+        message: resetPasswordMessages.passwordReseted,
+      })
+    );
+    yield put(push(LOGIN));
+  } catch (error) {
+    yield put(stopAction(type));
+    if (error.status === 422) {
+      yield call(setErrors, parseApiErrorsToFormik(error.data.errors));
+    }
+    yield put(resetPasswordError());
+  } finally {
+    yield put(stopAction(type));
+  }
+}
+
 export default function* appSaga() {
   yield takeLatest(LOGIN_REQUEST, authorize);
   yield takeLatest(FETCH_AUTHENTICATED_USER_REQUEST, fetchUser);
   yield takeLatest(LOGOUT_REQUEST, logout);
   yield takeLatest(FORGOT_PASSWORD_REQUEST, forgotPassword);
   yield takeLatest(REGISTER_REQUEST, register);
+  yield takeLatest(RESET_PASSWORD_REQUEST, resetPassword);
 }
